Conflicts:
      tests/test_conftest.py
      tox.ini

+++ tests/integration/testprocess.py
@@ -19,11 +19,12 @@
 
 """Base class for a subprocess run for tests.."""
 
-import sys
-import os.path
+import re
+import fnmatch
 
 import pytestqt.plugin  # pylint: disable=import-error
-from PyQt5.QtCore import pyqtSlot, pyqtSignal, QProcess, QObject
+from PyQt5.QtCore import pyqtSlot, pyqtSignal, QProcess, QObject, QElapsedTimer
+from PyQt5.QtTest import QSignalSpy
 
 
 class InvalidLine(Exception):
@@ -33,6 +34,18 @@ class InvalidLine(Exception):
     pass
 
 
+class ProcessExited(Exception):
+
+    """Raised when the child process did exit."""
+
+    pass
+
+
+class WaitForTimeout(Exception):
+
+    """Raised when wait_for didn't get the expected message."""
+
+
 class Process(QObject):
 
     """Abstraction over a running test subprocess process.
@@ -40,15 +53,15 @@ class Process(QObject):
     Reads the log from its stdout and parses it.
 
     Signals:
+        ready: Emitted when the server finished starting up.
         new_data: Emitted when a new line was parsed.
     """
 
-    PROCESS_NAME = None
+    ready = pyqtSignal()
     new_data = pyqtSignal(object)
 
     def __init__(self, parent=None):
         super().__init__(parent)
-        assert self.PROCESS_NAME is not None
         self._invalid = False
         self._data = []
         self.proc = QProcess()
@@ -63,7 +76,7 @@ class Process(QObject):
         raise NotImplementedError
 
     def _executable_args(self):
-        """Get the arguments to pass to the executable."""
+        """Get the executable and arguments to pass to it as a tuple."""
         raise NotImplementedError
 
     def _get_data(self):
@@ -78,13 +91,25 @@ class Process(QObject):
         self.read_log()
         return self._data
 
+    def _wait_signal(self, signal, timeout=5000, raising=True):
+        """Wait for a signal to be emitted.
+
+        Should be used in a contextmanager.
+        """
+        blocker = pytestqt.plugin.SignalBlocker(
+            timeout=timeout, raising=raising)
+        blocker.connect(signal)
+        return blocker
+
     @pyqtSlot()
     def read_log(self):
         """Read the log from the process' stdout."""
+        if not hasattr(self, 'proc'):
+            # I have no idea how this happens, but it does...
+            return
         while self.proc.canReadLine():
             line = self.proc.readLine()
-            line = bytes(line).decode('utf-8').rstrip('\r\n')
-            print(line)
+            line = bytes(line).decode('utf-8', errors='ignore').rstrip('\r\n')
 
             try:
                 parsed = self._parse_line(line)
@@ -93,33 +118,23 @@ class Process(QObject):
                 print("INVALID: {}".format(line))
                 continue
 
-            print('parsed: {}'.format(parsed))
             if parsed is not None:
                 self._data.append(parsed)
                 self.new_data.emit(parsed)
 
     def start(self):
         """Start the process and wait until it started."""
-        blocker = pytestqt.plugin.SignalBlocker(timeout=5000, raising=True)
-        blocker.connect(self.ready)
-        with blocker:
+        with self._wait_signal(self.ready, timeout=20000):
             self._start()
 
     def _start(self):
         """Actually start the process."""
-        if hasattr(sys, 'frozen'):
-            executable = os.path.join(os.path.dirname(sys.executable),
-                                      self.PROCESS_NAME)
-            args = []
-        else:
-            executable = sys.executable
-            args = [os.path.join(os.path.dirname(__file__),
-                                 self.PROCESS_NAME + '.py')]
-
-        self.proc.start(executable, args + self._executable_args())
+        executable, args = self._executable_args()
+        self.proc.readyRead.connect(self.read_log)
+        self.proc.start(executable, args)
         ok = self.proc.waitForStarted()
         assert ok
-        self.proc.readyRead.connect(self.read_log)
+        assert self.is_running()
 
     def after_test(self):
         """Clean up data after each test.
@@ -131,7 +146,70 @@ class Process(QObject):
         if self._invalid:
             raise InvalidLine
 
-    def cleanup(self):
+    def terminate(self):
         """Clean up and shut down the process."""
+        if not self.is_running():
+            raise ProcessExited
         self.proc.terminate()
         self.proc.waitForFinished()
+
+    def is_running(self):
+        """Check if the process is currently running."""
+        return self.proc.state() == QProcess.Running
+
+    def wait_for(self, timeout=5000, **kwargs):
+        """Wait until a given value is found in the data.
+
+        Keyword arguments to this function get interpreted as attributes of the
+        searched data. Every given argument is treated as a pattern which
+        the attribute has to match against.
+
+        The behavior of this method is slightly different depending on the
+        types of the filtered values:
+
+        - If the value is a string or bytes object and the expected value is
+          too, the pattern is treated as a fnmatch glob pattern.
+        - If the value is a string or bytes object and the expected value is a
+          compiled regex, it is used for matching.
+        - If the value is any other type, == is used.
+
+        Return:
+            The matched line.
+        """
+
+        # FIXME make this a context manager which inserts a marker in
+        # self._data in __enter__ and checks if the signal already did arrive
+        # after marker in __exit__, and if not, waits?
+
+        regex_type = type(re.compile(''))
+
+        spy = QSignalSpy(self.new_data)
+        elapsed_timer = QElapsedTimer()
+        elapsed_timer.start()
+
+        while True:
+            got_signal = spy.wait(timeout)
+            if not got_signal or elapsed_timer.hasExpired(timeout):
+                raise WaitForTimeout
+
+            for args in spy:
+                assert len(args) == 1
+                line = args[0]
+
+                matches = []
+
+                for key, expected in kwargs.items():
+                    if expected is None:
+                        continue
+
+                    value = getattr(line, key)
+
+                    if isinstance(expected, regex_type):
+                        matches.append(expected.match(value))
+                    elif isinstance(value, (bytes, str)):
+                        matches.append(fnmatch.fnmatchcase(value, expected))
+                    else:
+                        matches.append(value == expected)
+
+                if all(matches):
+                    return line
